Updated card tests.
